apps/backend/pyproject.toml
[project]
name = "fairy-demo-backend"
version = "0.1.0"
description = "Fairy web demo backend (FastAPI) for the research agent pipeline"
requires-python = ">=3.11,<3.14"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.0.0",
  "python-dotenv>=1.0.0",
]

[tool.uv]
dev-dependencies = [
  "ruff>=0.6.1",
]


